move package game
