+++ src/main/java/com/fonowizja/ox/game_elements/Board.java
@@ -6,7 +6,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicBoolean;
 
-import com.fonowizja.ox.game_elements.game.BoardObservator;
+import com.fonowizja.ox.game.BoardObservator;
 import lombok.AccessLevel;
 import lombok.Getter;
